// generators/src/types.ts
export type KarabinerRules = {
    description?: string;
    manipulators?: Manipulator[];
};

export type Manipulator = {
    description?: string;
    type: "basic";
    from: From;
    to?: To[];
    to_after_key_up?: To[];
    to_if_alone?: To[];
    parameters?: Parameters;
    conditions?: Conditions[];
    to_delayed_action?: {
        to_if_invoked: [
            { set_variable: { name: string; value: number } },
            { key_code: KeyCode },
        ];
    };
};

export type Parameters = Partial<{
    "basic.simultaneous_threshold_milliseconds": number;
    "basic.to_delayed_action_delay_milliseconds": number;
}>;

type Conditions =
    | FrontMostApplicationCondition
    | DeviceCondition
    | KeybaordTypeCondition
    | InputSourceCondition
    | VaribaleCondition
    | EventChangedCondition;
// ...
export type From = {
    key_code?: KeyCode;
    simultaneous?: SimultaneousFrom[];
    simultaneous_options?: SimultaneousOptions;
    modifiers?: Modifiers;
};

export type Modifiers = {
    optional?: string[];
    mandatory?: string[];
};

export type To = {
    key_code?: KeyCode;
    modifiers?: KeyCode[];
    shell_command?: string;
    set_variable?: {
        name: string;
        value: boolean | number | string;
    };
    mouse_key?: MouseKey;
    pointing_button?: string;
    /**
     * Power Management plugin
     * @example: sleep system
     * @see: {@link https://karabiner-elements.pqrs.org/docs/json/complex-modifications-manipulator-definition/to/software_function/iokit_power_management_sleep_system/}
     */
    software_function?: SoftwareFunction;
};